backend/requirements.txt
torch~=2.6.0
ultralytics~=8.3.73
roboflow~=1.1.53
python-dotenv~=1.0.1
Flask==3.1.0
Flask-Cors==5.0.0
opencv-python==4.10.0.84
